+++ web/menugen-web/src/App.tsx
@@ -12,6 +12,7 @@ import { MenuPage }          from './pages/Menu/MenuPage';
 import { FamilyPage }        from './pages/Family/FamilyPage';
 import { DiaryPage }         from './pages/Diary/DiaryPage';
 import { FridgePage }        from './pages/Fridge/FridgePage';
+import { ShoppingPage }      from './pages/Shopping/ShoppingPage';
 import { SubscriptionsPage } from './pages/Subscriptions/SubscriptionsPage';
 import { ProfilePage }       from './pages/Profile/ProfilePage';
 import { SpecialistDashboardPage } from './pages/specialist/SpecialistDashboardPage';
@@ -44,6 +45,7 @@ const AppRoutes: React.FC = () => {
         <Route path="family"        element={<FamilyPage />} />
         <Route path="diary"         element={<DiaryPage />} />
         <Route path="fridge"        element={<FridgePage />} />
+        <Route path="shopping"      element={<ShoppingPage />} />
         <Route path="subscriptions" element={<SubscriptionsPage />} />
         <Route path="profile"       element={<ProfilePage />} />
         <Route path="specialist"                                          element={<SpecialistDashboardPage />} />